requirements.txt
streamlit==1.39.0
tensorflow==2.17.0
numpy>=1.26.0,<2.0.0
matplotlib==3.7.5
Pillow==9.5.0
setuptools>=67.0.0